Current weather started
I started  with the current temperature and weather conditions. Connecting the elements across documents.
I like the words to be capitalized and the numbers to be rounded so I'm making sure everything is been done correctly and that it shows the right things.
I started working on the time and date but I'm yet to do some research so I'll continue that later