Forecast API
I decided to remove the forecast HTML and add it with Javascript as a loop for each day instead. It makes the code simpler and more readable.
API and each day element are now linked and showing the right data.
I had to also modify some of the CSS so it applies to changes made to the rest of the code.